Dodi RePacker banned due to lies, ads, money & more lies + excuses
Background:
https://redd.it/hi7gb6

This is not the first time, there are repacks with ads even malware and I take action now to protect people using this clown repacks. His page are riddled with ads, false information and there are a lot of lies.

About this commit:
This commit is NOT to roast his work, in fact I respect some of his work (not all) but doing Warez for money is a no-go and it's an undeniable fact that money is what he ultimately want, he does it not for fun or to help the scene, his main intention always was to get a personal benefit out of it.

I do believe we have better repackers who doing this actually only to help and this is what this repo is about. List people/groups who do it for the scene not to gain something out of it. Donations are of course no problem because they are optional and to show respect and appreciation. 

This been said, I will take a closer look at all repackers and check their background more frequently. 

Please do not roast this group or individual persons now or make a drama out of it, just ignore and block such pages, that's the best you can do to show your disrespect. Talking about them only helps them to get more attention which is the opposite such groups/sites/people deserve.