Copy To GDrive "App" (script added)
This app provides the capability to copy a Google Drive folder. It will copy all contents of the folder and preserve the internal structure of the files and folders. Optionally, you can also copy any sharing permissions over to the new folder.

If a folder is nested within other folders of your Google Drive, it is often convenient to make a new copy in the same location. This is the default behavior. However, if you would rather have the copy appear at the top level of your Google Drive, you can select to place your copy at the Root level.